requirements.txt
camel-ai 
PyMuPDF
requests 
python-dotenv
gradio
google-generativeai